integrated_project/backend/requirements.txt

flask
flask-cors
python-dotenv
PyPDF2
langchain
langchain-community
langchain-google-genai
google-generativeai
